broke tests into smaller ones
